initial compile based bump
This shows the naive difference between the pip install and the pip
compile based requirements compilation methods. After this I'll add on a
few commits to reduce the diff to only non-functional changes (eg line
order). Some of the later commits may undo beneficial behaviour in
favour of compatibility with the existing pip install based method.