tab-give --recursive without direct tree manipulation
The existing `_tree_tab_give()` method manipulates the tree manually and
has a two setp process to create tabs and then position them.
I wanted to see if I could create and position tabs in a single step,
assuming it would be simpler.

This did not turn out as simple as I had hoped!
It was hoping I could iterate through the tabs to be given and open new
tabs in the destination window in the right order, then they would end
up in the right place and we would be done.

The complication is that I don't think there's any way to traverse the
tree and open new tabs without doing extra work to position tabs. In
this case we have to figure out when we go up and down levels in the
tree, so we can pass the right args to `tabopen()` and have the right
tab focused to have the new one opened relative to. Going down the tree
is pretty easy, going up the tree we have to figure out how far we went,
also not so bad as it turns out.

On balance I'm not sure this is a clear win. The new function is more
lines of code and needs heuristic to understand it's position in the tree
while traversing. On the plus side the new code maintains the collapsed
state of subtrees! For the previous code I tried to add that but it
seemed to blow up on the sanity check in update_tab_titles where it's
gets indexes that are out of range of the visible tabs.