Fix `tab-move end` for tree tabs
Previously moving a tab to a higher tab number ended up putting it right
before the target node. Since the target node stayed in the same place that
actually meant that the tab being moved didn't make it to the indicated index.
The `new_idx_relative += 1` bit in this change accounts for that.

I also somewhat re-wrote the block of code to read a bit more clearly to me.
Previously it called nodes "tabs", and it looked up `tabs[new_idx]` twice.
It made it a bit difficult to spot logic issues.

What this block of code is doing is:
1. find target node by indexing into the list of tabs as per the tab bar
2. remove the current node from the tree
3. insert the current node into the sibling list of the target node

It's important that we look up the target node before removing the current.
Otherwise everything could shift across and we could end up putting the tab in
the next subtree.

Also added a bunch of tests that let me make sure I was making changes
correctly.