Add common "ensure tree tab" method in commands.py
There was a few checks of `is_treetabbedbrowser` and repeated errors messages.
I wanted to move them to a common method, but might have overestimated how
many places could use it. There are several tree tab specific arguments that
aren't treated as fatal, for example `tab-give --recursive` and `tab-prev
--sibling`. I'm not sure if they should be fatal or not. There is some value
in defaulting to the non-tree tab behavior if someone wants to use the same
bindings across tree tab and non-tree tab basedirs. There is also value in
being clear that what you've asked for can't be delivered.

If indeed there is only one or two places where we need to complain about an
argument, and not a whole command, then we can moved the magic command name
finding bit into the Command object and just hardcode the command name in the
two places where we want to warn about a specific argument.

There is also the `tree-tab-create-group` command which doesn't bother
checking that yo have a tree tab enabled windows because it doesn't actually
depend on any tree tab features, it's just an alternate to about:blank.