create dictionary dir in test_init
The `test_existing_dict` test in `test_webenginesettings.py` is failing
because spell checking is no enabled on the profile when we expect it to
be.
It's failing since https://bugreports.qt.io/browse/QTBUG-124110 ensured
that spell checking was disabled if webengine couldn't find a
dictionaries directory.
Unfortunately, the check for dictionary directories is done when
webengine is initialized and then saved to a static variable. So for
these unit tests the check introduced in that QTBUG ticket is going to
be looking at the resources directory of the first test for every
subsequent one.

I can't figure out a way to re-initialize this static stuff in some way,
potentially some of these spell check tests will have to move to end2end
tests. The tests that need to read files work fine because they work on
our logic, the one that is failing just needs to get past the "does a
directory exist" check. So for now I'm just creating the dictionary
directory in the resources dir for the test that the "not found" warning
messages are pointing to. That way subsequent tests will work, for now.

The static cachgin: https://github.com/qt/qtwebengine/blob/6.8.0/src/core/web_engine_library_info.cpp#L248

Ref: https://github.com/qutebrowser/qutebrowser/issues/8242#issuecomment-2333609589